backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Async & Tasks
celery==5.3.6
redis==5.0.1

# Storage
minio==7.2.3

# AI/ML
numpy<2.0.0
sentence-transformers==2.3.1
chromadb==0.4.22
httpx==0.26.0

# Document Generation
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Payment
razorpay==1.4.1
stripe==7.11.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
